Simplify helper functions.
- Extract `is_local_testing` shared by corrdiff_datagen.py and era5.py
- Move `print_slices_over_time` into zarr_dump.py
- Use identical REF grid path for both local testing and BIG remote.
